requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
load-dotenv==0.1.0
multidict==6.0.4
openai==0.26.1
praw==7.6.1
prawcore==2.3.0
pycodestyle==2.10.0
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
requests==2.28.2
telegram==0.0.1
tomli==2.0.1
tqdm==4.64.1
update-checker==0.18.0
urllib3==1.26.14
websocket-client==1.5.0
yarl==1.8.2
